+++ lib/python/ensembl/microbes/auxiliary_files/PHIbase2/PHI_DUMMY.obo
