+++ lib/python/ensembl/microbes/auxiliary_files/PHIbase2/createP2PDB.sql
