--- lib/python/ensembl/microbes/runnable/__init__.py
@@ -1,9 +1,0 @@
-#Ensembl
-#Copyright [1999-2015] Wellcome Trust Sanger Institute and the EMBL-European Bioinformatics Institute
-#Copyright [2016-2021] EMBL-European Bioinformatics Institute
-#
-#This product includes software developed at:
-#
-#EMBL-European Bioinformatics Institute
-#Wellcome Trust Sanger Institute
-
