+++ lib/python/ensembl/microbes/runnables/PHIbase_2/EnsemblCoreReader.py
@@ -1,13 +1,17 @@
-#Ensembl
-#Copyright [1999-2015] Wellcome Trust Sanger Institute and the EMBL-European Bioinformatics Institute
-#Copyright [2016-2021] EMBL-European Bioinformatics Institute
-#
-#This product includes software developed at:
-#
-#EMBL-European Bioinformatics Institute
-#Wellcome Trust Sanger Institute
 
-# standaloneJob.pl eHive.examples.TestRunnable -language python3
+# See the NOTICE file distributed with this work for additional information
+# regarding copyright ownership.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 import os
 import subprocess
@@ -18,8 +22,8 @@ import pymysql
 import eHive
 import datetime
 import re
-import ensembl.microbes.runnable.PHIbase_2.core_DB_models as core_db_models
-import ensembl.microbes.runnable.PHIbase_2.interaction_DB_models as interaction_db_models
+import models as core_db_models
+import ensembl.microbes.auxiliary_files.PHIbase2.interaction_DB_models as interaction_db_models
 import requests
 from xml.etree import ElementTree
 import time
