+++ lib/python/ensembl/microbes/runnables/PHIbase_2/InteractionKeys.py
@@ -1,13 +1,16 @@
-#Ensembl
-#Copyright [1999-2015] Wellcome Trust Sanger Institute and the EMBL-European Bioinformatics Institute
-#Copyright [2016-2021] EMBL-European Bioinformatics Institute
+# See the NOTICE file distributed with this work for additional information
+# regarding copyright ownership.
 #
-#This product includes software developed at:
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# http://www.apache.org/licenses/LICENSE-2.0
 #
-#EMBL-European Bioinformatics Institute
-#Wellcome Trust Sanger Institute
-
-# standaloneJob.pl eHive.examples.TestRunnable -language python3
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 import os
 import subprocess
@@ -15,8 +18,8 @@ import sqlalchemy as db
 import sqlalchemy_utils as db_utils
 import pymysql
 import eHive
-import ensembl.microbes.runnable.PHIbase_2.interaction_DB_models as interaction_db_models
-import ColumnMapper as col_map
+import ensembl.microbes.auxiliary_files.PHIbase2.interaction_DB_models as interaction_db_models
+import ensembl.microbes.auxiliary_files.PHIbase2.ColumnMapper as col_map
 import csv
 
 from xml.etree import ElementTree
