+++ lib/python/ensembl/microbes/runnables/PHIbase_2/InteractorDataManager.py
@@ -1,31 +1,38 @@
-#Ensembl
-#Copyright [1999-2015] Wellcome Trust Sanger Institute and the EMBL-European Bioinformatics Institute
-#Copyright [2016-2021] EMBL-European Bioinformatics Institute
+# See the NOTICE file distributed with this work for additional information
+# regarding copyright ownership.
 #
-#This product includes software developed at:
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# http://www.apache.org/licenses/LICENSE-2.0
 #
-#EMBL-European Bioinformatics Institute
-#Wellcome Trust Sanger Institute
-
-# standaloneJob.pl eHive.examples.TestRunnable -language python3
-
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+\
 import os
 import subprocess
 import eHive
 
-class InteractionTable(eHive.BaseRunnable):
-    """Manages the data that goes into the Interaction table"""
+class InteractorDataManager(eHive.BaseRunnable):
+    """Makes sure all fields related to the interactors are gathered before attempting to write to mysql DB"""
 
     def param_defaults(self):
         return { }
 
     def fetch_input(self):
+        self.warning("Fetch InteractorDataManager")
         self.param('failed_job', '')
         phi_id = self.param_required('PHI_id')
-        #self.check_param('patho_ensembl_gene_stable_id')
+        self.check_param('patho_ensembl_gene_stable_id')
+        self.check_param('patho_molecular_structure')
+        self.check_param('host_ensembl_gene_stable_id')
+        self.check_param('host_molecular_structure')   
 
     def run(self):
-        self.warning("InteractionTable run")
+        self.warning("InteractorDataManager run")
         self.get_interactor_fields()
         
     def get_interactor_fields(self):
@@ -64,7 +71,6 @@ class InteractionTable(eHive.BaseRunnable):
             entries_list = self.build_output_hash()
             for entry in entries_list:
                 self.dataflow(entry, 1)
-            print(f"{phi_id} written to DBwriter")
         else:
             #output_hash = [{"uncomplete_entry": self.param('failed_job')} ]
             print(f"{phi_id} written to FailedJob")
