+++ lib/python/ensembl/microbes/runnables/PHIbase_2/TestRunnable.py
@@ -14,8 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# standaloneJob.pl eHive.examples.TestRunnable -language python3
-
+`cat licence.txt`
 import os
 import subprocess
 import unittest
